+++ src/vprint_hex.c
@@ -68,7 +68,7 @@ int	printp(va_list args)
 	int		j;
 
 	p = va_arg(args, void *);
-	str = ultoa((unsigned long)p);
+	str = ultoa((unsigned long long)p);
 	if (!str)
 		return (-1);
 	j = pprintf(str);
